+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 2.8.0
+- Change: `InPageMini` and `InPageStandard` now use `NotoSans` font instead of `NotoSansCJK` for Japanese and Korean languages
+- Refactor: Reduce SDK binary size by using minimal-required subset of `NotoSans` fonts
+- Feature: Validate fonts support all the localization texts during PR checks
+
 ### 2.7.5
 * Fix: Password Reset button doesn’t work
 * Fix: After the login is succeeded, you can never go back to the app automatically